1.5.3
1. 版本版本: 1.5.3

  2. 更新时间: 2017-08-10 11:54:38

  3. 更新内容:

    * 修改 pod 警告信息，Demo 版本信息

+++ SwpCateGory/UIKit/UIColorCateGory/UIColor+SwpColor.m
@@ -241,9 +241,9 @@
 
 + (NSString*)toStrByUIColor1:(NSInteger)hexadecimalValue {
     
-    NSLog(@"%ld", (hexadecimalValue & 0xFF0000) >> 16);
-    NSLog(@"%ld", (hexadecimalValue & 0xFF00) >> 8);
-    NSLog(@"%ld", (hexadecimalValue & 0xFF));
+    NSLog(@"%ld", (long)(hexadecimalValue & 0xFF0000) >> 16);
+    NSLog(@"%ld", (long)(hexadecimalValue & 0xFF00) >> 8);
+    NSLog(@"%ld", (long)(hexadecimalValue & 0xFF));
     return nil;
 }
